requirements.txt
opencv-contrib-python>=4.5.0
numpy>=1.19.0
imutils>=0.5.4
pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.02